+++ test/html/runtest.sh
@@ -1,0 +1,1 @@
+npx playwright test /test/html/test.spec.js